refactor: Renamed option_types.dart
