+++ lib/src/config/config.dart
@@ -5,6 +5,6 @@ export 'configuration_parser.dart';
 export 'configuration.dart';
 export 'file_system_options.dart';
 export 'multi_config_source.dart';
-export 'options.dart';
+export 'option_types.dart';
 export 'option_groups.dart';
 export 'source_type.dart';
